refactor: revise FAQ page
